pyproject.toml
[project]
name = "example"
version = "0.1.0"
description = ""
authors = [
  { name = "duanluan", email = "duanluan@example.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9,<=3.13"
dependencies = [
  "pyyaml>=6.0.3",
  "py7zr>=1.0.0",
  "apscheduler>=3.11.0",
  "pyinstaller>=6.16.0",
  "pydantic>=2.11.9",
  "pydantic-settings>=2.11.0"
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
